Move loading of configuration back out to Main
So that Publisher class is more isolated from external state

Also tidy up constructor vs main method of Publisher

PR feedback
